+++ tests/Reddit.NET.Client.UnitTests/Authentication/AuthenticatorTests.cs
@@ -3,6 +3,8 @@ using System.Diagnostics.CodeAnalysis;
 using System.Net;
 using System.Net.Http;
 using System.Text.Json;
+using System.Threading;
+using System.Threading.RateLimiting;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
@@ -11,7 +13,6 @@ using Reddit.NET.Client.Authentication;
 using Reddit.NET.Client.Authentication.Context;
 using Reddit.NET.Client.Authentication.Credential;
 using Reddit.NET.Client.Command;
-using Reddit.NET.Client.Command.RateLimiting;
 using Reddit.NET.Client.Models.Internal;
 using Reddit.NET.Client.UnitTests.Shared;
 
@@ -22,7 +23,7 @@ namespace Reddit.NET.Client.UnitTests.Authentication
     {
         private ILogger<CommandExecutor> _logger;
         private IHttpClientFactory _httpClientFactory;
-        private IRateLimiter _rateLimiter;
+        private RateLimiter _rateLimiter;
         private MockHttpMessageHandler _httpMessageHandler;
         private CommandExecutor _commandExecutor;
 
@@ -31,7 +32,7 @@ namespace Reddit.NET.Client.UnitTests.Authentication
         {
             _logger = Substitute.For<ILogger<CommandExecutor>>();
             _httpClientFactory = Substitute.For<IHttpClientFactory>();
-            _rateLimiter = Substitute.For<IRateLimiter>();
+            _rateLimiter = Substitute.For<RateLimiter>();
             _httpMessageHandler = new MockHttpMessageHandler();
 
             _httpMessageHandler.RequestFunc = (request) =>
@@ -44,7 +45,7 @@ namespace Reddit.NET.Client.UnitTests.Authentication
             var successfulLease = SuccessfulLease();
 
             _rateLimiter
-                .AcquireAsync(Arg.Any<int>())
+                .AcquireAsync(Arg.Any<int>(), Arg.Any<CancellationToken>())
                 .Returns(successfulLease);
 
             _commandExecutor = new CommandExecutor(
@@ -632,9 +633,9 @@ namespace Reddit.NET.Client.UnitTests.Authentication
             Assert.AreEqual("grant_type=refresh_token&refresh_token=...&duration=permanent", requestContent);
         }
 
-        private static ValueTask<PermitLease> SuccessfulLease()
+        private static ValueTask<RateLimitLease> SuccessfulLease()
         {
-            var lease = Substitute.For<PermitLease>();
+            var lease = Substitute.For<RateLimitLease>();
 
             lease.IsAcquired.Returns(true);