+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased - 2025-04-18
+
+- Upgrade to .NET 10
+- Deprecate ability to award user content, due to Reddit API removal
+
 ## [2.3.0] - 2022-11-04
 
 - Add ability to load unresolved comment threads